requirements.txt
anyio==3.6.1
asgiref==3.5.2
attrs==21.4.0
boto3==1.24.6
botocore==1.27.6
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
coverage==6.4.1
databases==0.6.0
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fastapi==0.78.0
future==0.18.2
fuzzywuzzy==0.18.0
geographiclib==1.52
geopy==2.2.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
jmespath==1.0.0
packaging==21.3
passlib==1.7.4
peewee==3.14.10
pluggy==1.0.0
presto-python-client==0.8.2
psycopg2-binary==2.8.4
py==1.11.0
pyasn1==0.4.8
pydantic==1.9.1
pydantic[email]
PyHive==0.6.5
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-Levenshtein==0.12.2
python-multipart==0.0.5
requests==2.28.0
rsa==4.8
s3transfer==0.6.0
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.19.1
tenacity==8.0.1
tomli==2.0.1
typing_extensions==4.2.0
Unidecode==1.3.4
urllib3==1.26.9
uvicorn==0.17.6
watchtower==3.0.0
